refactor dpm proof pack ai route
